fix for migration integration test failure

+++ src/Microsoft.Health.Dicom.SqlServer/Features/Schema/Sql/Scripts/Sequences.sql
@@ -61,7 +61,7 @@ CREATE SEQUENCE dbo.WorkitemKeySequence
     CACHE 10000
 
 CREATE SEQUENCE dbo.WorkitemWatermarkSequence
-    AS INT
+    AS BIGINT
     START WITH 1
     INCREMENT BY 1
     MINVALUE 1
